==> 1 Parquet to CSV.ipynb <==
import pandas as pd
import os
import string

files_df = pd.DataFrame(columns=["FilePath"])

directory = "bbc_news_alltime"

for root, dirs, files in os.walk(directory):
    for file in files:
        if file.endswith(".parquet"):
            file_path = os.path.join(root, file)
            files_df = pd.concat([files_df, pd.DataFrame({"FilePath": [file_path]})], ignore_index=True)
            
print(files_df)

full_df = pd.DataFrame([])

for index, file in files_df.iterrows():
    article = pd.read_parquet(file["FilePath"])
    full_df = pd.concat([full_df, article], ignore_index=True)

full_df = full_df.sort_values(by="published_date", ascending=True).reset_index(drop=True)

def clean_text_translate(text):
    if not text:
        return None
    text = text.translate(str.maketrans("", "", string.punctuation))
    return text.replace("\n", " ").replace("\r", " ")

full_df['content'] = full_df['content'].apply(clean_text_translate)
full_df['description'] = full_df['description'].apply(clean_text_translate)

full_df

full_df.to_csv("bbc_full.csv", index=False)
