==> 1 SSIM Scrape.ipynb <==
import pandas as pd
from datetime import datetime, timedelta
from tqdm import tqdm
from selenium import webdriver
from selenium.webdriver.common.by import By
import random

def generate_archive_links(start_date, end_date):
    archive_links = []
    
    current_date = start_date
    while current_date <= end_date:
        year_month = current_date.strftime('%Y-%m')
        day = current_date.strftime('%d')
        archive_links.append(f"https://www.reuters.com/archive/{year_month}/{day}/1")
        current_date += timedelta(days=1)  # Move to the next day

    return archive_links

start_date = datetime(2006, 5, 22)
end_date = datetime(2023, 10, 31)

generated_links = generate_archive_links(start_date, end_date)

generated_links

proxy_list = [
"http://49.12.235.70:8081",
"http://80.67.8.6:80",
"http://5.252.22.45:80",
"http://198.49.68.80:80",
"http://62.92.103.236:8080",
"http://20.13.148.109:8080",
"http://47.74.152.29:8888",
"http://41.204.53.26:80",
"http://114.218.165.11:8089",
"http://67.43.227.227:9797",
"http://38.156.73.148:8080",
"http://123.30.154.171:7777",
"http://121.227.179.178:8089",
"http://143.64.8.21:8080",
"http://109.236.83.153:8888",
"http://183.240.46.42:80",
"http://79.174.12.190:80",
"http://67.43.236.20:15549",
"http://115.231.181.40:8128",
"http://36.64.68.169:8789",
"http://84.201.151.210:80",
"http://170.106.183.248:8080",
"http://67.43.227.227:25507",
"http://121.227.179.144:8089",
"http://67.43.227.227:11875",
"http://185.159.153.234:80",
"http://72.10.160.93:9475",
"http://177.93.40.182:999",
"http://185.44.65.187:9595",
"http://72.10.164.178:6547",
"http://36.92.193.189:80",
"http://67.43.227.227:30017",
"http://73.117.183.115:80",
"http://134.209.29.120:3128",
"http://58.20.248.139:9002",
"http://145.40.97.148:9400",
"http://172.94.32.48:80",
"http://196.1.95.124:80",
"http://154.0.12.163:80",
"http://114.156.77.107:8080",
"http://23.247.136.245:80",
"http://103.133.25.57:3125",
"http://154.90.55.37:80",
"http://103.73.66.36:8085",
"http://125.124.225.211:3128",
"http://67.43.227.226:19637",
"http://189.22.234.47:80",
"http://72.10.160.93:26841",
"http://43.134.68.153:3128",
"http://121.227.179.166:8089",
"http://95.216.208.103:8118",
"http://72.10.164.178:5811",
"http://72.10.164.178:15501",
"http://58.209.137.200:8089",
"http://67.43.228.253:12777",
"http://103.127.1.130:80",
"http://181.12.169.91:11201",
"http://189.22.234.41:80",
"http://72.10.160.170:21047",
"http://51.15.241.34:3128",
"http://115.72.36.85:10005",
"http://67.43.236.19:17303",
"http://72.10.164.178:26417",
"http://67.43.228.253:18547",
"http://67.43.236.18:16037",
"http://67.43.236.22:28935",
"http://41.204.53.19:80",
"http://41.173.24.38:80",
"http://134.249.185.223:41890",
"http://67.43.227.230:3559",
"http://67.43.227.227:30785",
"http://190.110.226.122:80",
"http://67.43.227.227:12163",
"http://67.43.236.19:15477",
"http://82.146.37.145:80",
"http://181.209.82.206:999",
"http://63.143.57.117:80",
"http://128.199.202.122:8080",
"http://45.5.119.222:999",
"http://114.218.165.9:8089",
"http://47.91.104.88:3128",
"http://103.125.174.13:7777",
"http://62.210.215.36:80",
"http://45.70.201.253:999",
"http://20.106.146.212:5001",
"http://198.199.86.11:8080",
"http://114.218.150.36:8089",
"http://72.10.160.94:7295",
"http://27.189.129.57:8089",
"http://72.10.160.170:33169",
"http://189.22.234.38:80",
"http://67.43.236.19:20217",
"http://196.251.223.29:8104",
"http://39.109.113.97:4090",
"http://179.60.240.69:53281",
"http://145.40.97.148:10004",
"http://43.133.59.220:3128",
"http://103.189.123.110:8080",
"http://103.215.187.67:8080",
"http://62.33.207.202:80",
"http://145.40.97.148:80",
"http://218.91.69.252:8888",
"http://178.48.68.61:18080",
"http://159.89.208.15:4750",
"http://123.126.158.50:80",
"http://181.78.108.225:999",
"http://122.175.19.164:80",
"http://202.191.125.147:8080",
"http://189.43.42.111:80",
"http://128.199.202.122:3128",
"http://107.189.8.240:8080",
"http://190.95.132.187:999",
"http://116.203.50.3:3128",
"http://222.122.110.26:80",
"http://171.248.158.200:3127",
"http://67.43.228.254:25897",
"http://175.139.233.76:80",
"http://38.156.72.57:8080",
"http://143.42.66.91:80",
"http://72.10.164.178:26401",
"http://58.209.137.245:8089",
"http://4.175.200.138:8080",
"http://154.85.58.149:80",
"http://67.43.227.228:19253",
"http://121.236.236.205:8089",
"http://180.191.42.241:8082",
"http://67.43.228.254:25705",
"http://34.122.187.196:80",
"http://113.192.31.5:1111",
"http://111.59.4.88:9002",
"http://114.218.165.34:8089",
"http://78.28.152.111:80",
"http://118.97.164.19:8080",
"http://221.6.139.190:9002",
"http://103.181.92.250:82",
"http://188.132.222.132:8080",
"http://181.236.247.8:8807",
"http://36.90.48.207:8080",
"http://185.191.236.162:3128",
"http://186.65.107.2:666",
"http://67.43.236.20:12457",
"http://68.178.168.41:80",
"http://129.226.193.16:3128",
"http://194.250.197.206:80",
"http://149.34.210.62:9090",
"http://109.61.42.223:80",
"http://72.10.164.178:10367",
"http://85.232.243.235:8080",
"http://31.211.82.158:3128",
"http://176.9.239.181:80",
"http://109.176.198.142:80",
"http://114.218.165.55:8089",
"http://72.10.160.171:31259",
"http://103.174.236.83:8080",
"http://89.117.130.19:80",
"http://67.43.227.227:15061",
"http://185.173.130.148:3128",
"http://200.34.227.28:8080",
"http://67.43.227.227:27863",
"http://168.119.141.135:80",
"http://70.186.128.126:8080",
"http://61.129.2.212:8080",
"http://189.7.145.111:3128",
"http://159.192.139.178:8080",
"http://195.114.209.50:80",
"http://67.43.227.229:30467",
"http://148.113.194.171:80",
"http://54.39.163.156:3128",
"http://131.196.42.95:667",
"http://67.43.227.226:9687",
"http://157.20.233.212:3125",
"http://91.222.238.112:80",
"http://46.161.194.242:8080",
"http://73.90.88.50:80",
"http://72.10.160.90:26289",
"http://67.43.227.227:12549",
"http://141.95.1.186:3128",
"http://103.112.66.7:80",
"http://185.89.156.2:44224",
"http://101.109.176.175:8080",
"http://62.84.245.79:80",
"http://67.43.227.227:4587",
"http://103.186.12.10:8080",
"http://168.119.53.93:80",
"http://47.91.121.127:9080",
"http://67.43.228.253:10669",
"http://194.147.33.5:8080",
"http://103.115.31.161:8090",
"http://114.216.205.236:8089",
"http://103.112.66.3:80",
"http://145.40.97.148:10003",
"http://37.211.84.100:8080",
"http://103.125.16.69:8080",
"http://121.227.183.160:8089",
"http://115.236.144.234:3128",
"http://38.180.209.139:80",
"http://47.88.18.204:8023",
"http://85.214.107.177:80",
"http://5.161.103.41:88",
"http://118.99.103.200:8080",
"http://114.9.27.142:8080",
"http://67.43.236.20:5543",
"http://115.178.48.138:8080",
"http://27.147.139.225:58080",
"http://36.92.60.234:8080",
"http://84.39.112.144:3128",
"http://51.91.97.136:3128",
"http://85.119.120.55:8080",
"http://20.233.44.207:80",
"http://83.222.7.245:8080",
"http://72.10.160.173:21965",
"http://194.104.136.243:3128",
"http://103.159.188.250:80",
"http://77.232.128.191:80",
"http://72.10.164.178:31269",
"http://103.156.16.227:8080",
"http://186.3.76.227:999",
"http://132.145.134.243:8000",
"http://31.40.248.2:8080",
"http://128.140.113.110:1080",
"http://183.100.14.134:8000",
"http://103.148.24.234:8080",
"http://67.43.227.227:11611",
"http://103.23.141.45:8181",
"http://196.1.92.12:80",
"http://186.113.232.105:8080",
"http://188.132.150.200:8080",
"http://80.249.112.162:80",
"http://103.156.248.55:8080",
"http://65.109.13.237:80",
"http://200.127.68.173:3128",
"http://67.43.228.254:6897",
"http://67.43.236.20:2305",
"http://181.209.81.82:999",
"http://103.63.26.226:1111",
"http://197.232.85.163:8080",
"http://188.132.221.170:8080",
"http://193.34.144.169:80",
"http://72.10.160.90:22539",
"http://67.43.227.230:10221",
"http://103.106.219.219:8080",
"http://103.186.90.41:8080",
"http://114.216.205.232:8089",
"http://38.156.72.40:8080",
"http://13.80.134.180:80",
"http://5.2.228.168:8888",
"http://123.30.154.38:2008",
"http://179.42.72.186:85",
"http://34.81.160.132:80",
"http://158.255.77.169:80",
"http://51.254.78.223:80",
"http://114.218.165.52:8089",
"http://45.76.122.178:80",
"http://51.89.255.67:80",
"http://36.95.155.74:3126",
"http://122.185.180.190:32650",
"http://67.43.227.226:30785",
"http://112.78.167.56:8080",
"http://43.129.238.55:8085",
"http://103.189.197.75:8181",
"http://72.10.164.178:30655",
"http://192.145.205.97:8080",
"http://67.43.228.253:13975",
"http://88.119.139.237:53281",
"http://58.246.58.150:9002",
"http://103.120.6.46:80",
"http://72.10.160.174:29821",
"http://49.145.181.132:8081",
"http://145.40.97.148:10005",
"http://194.186.248.97:80",
"http://180.191.59.254:8081",
"http://188.125.167.82:8080",
"http://190.110.34.24:999",
"http://49.12.235.70:80",
"http://189.43.42.105:80",
"http://103.101.125.18:28062",
"http://154.65.39.8:80",
"http://195.158.26.246:8080",
"http://51.210.54.186:80",
"http://113.192.30.183:8080",
"http://46.20.56.41:8080",
"http://43.243.141.11:8080",
"http://108.170.12.12:80",
"http://47.91.65.23:3128",
"http://43.134.118.223:3128",
"http://67.43.227.228:30711",
"http://123.58.199.17:8168",
"http://153.122.62.109:8118",
"http://67.43.228.254:19037",
"http://27.254.46.192:80",
"http://72.10.160.90:9475",
"http://4.155.2.13:9401",
"http://72.10.160.171:29161",
"http://67.43.227.227:13297",
"http://220.248.70.237:9002",
"http://23.82.137.162:80",
"http://62.210.15.199:80",
"http://103.76.150.74:8080",
"http://47.251.43.115:33333",
"http://103.255.222.1:80",
"http://191.101.80.162:80",
"http://23.237.145.36:31288",
"http://116.246.9.164:443",
"http://15.235.12.19:3128",
"http://103.168.44.209:3127",
"http://121.227.183.190:8089",
"http://1.20.203.149:8080",
"http://161.34.40.110:3128",
"http://72.10.160.90:13321",
"http://154.90.48.76:80",
"http://36.64.217.27:1313",
"http://162.223.90.130:80",
"http://179.48.11.6:8085",
"http://43.134.229.98:3128",
"http://103.156.75.41:8181",
"http://188.40.59.208:3128",
"http://189.22.234.40:80",
"http://201.77.128.1:8080",
"http://190.104.129.38:22800",
"http://157.100.55.143:999",
"http://138.185.34.123:80",
"http://47.242.47.64:8888",
"http://129.213.89.36:80",
"http://138.91.159.185:80",
"http://8.213.215.187:30000",
"http://189.43.42.96:80",
"http://72.10.164.178:10581",
"http://67.43.228.254:24675",
"http://144.126.216.57:80",
"http://115.79.27.106:1001",
"http://111.1.61.49:3128",
"http://95.177.184.57:80",
"http://115.72.37.233:10002",
"http://138.204.233.189:58080",
"http://138.0.141.46:8080",
"http://62.92.103.227:8080",
"http://159.65.245.255:80",
"http://103.177.199.57:1111",
"http://103.175.46.230:3125",
"http://67.43.228.253:19443",
"http://51.75.206.209:80",
"http://84.247.152.41:8000",
"http://90.85.169.218:3128",
"http://145.40.97.148:10006",
"http://222.124.202.144:8080",
"http://67.43.227.229:19253",
"http://170.64.175.174:8089",
"http://72.10.160.90:26519",
"http://103.147.246.78:8080",
"http://36.66.170.193:8080",
"http://199.195.253.14:1080",
"http://34.87.84.105:80",
"http://34.135.166.24:80",
"http://72.10.160.94:26289",
"http://186.96.67.37:999",
"http://45.188.111.242:999",
"http://47.88.59.79:82",
"http://67.43.227.227:3559",
"http://43.134.33.254:3128",
"http://8.213.151.128:3128",
"http://61.145.214.107:65533",
"http://63.143.57.116:80",
"http://41.32.177.42:3128",
"http://145.40.97.148:10001",
"http://189.22.234.44:80",
"http://67.43.227.229:24041",
"http://51.161.56.52:80",
"http://37.120.192.154:8080",
"http://103.49.202.250:80",
"http://112.78.47.188:8080",
"http://66.54.106.56:8103",
"http://223.197.178.186:3128",
"http://12.176.231.147:80",
"http://51.79.170.92:80",
"http://67.43.228.253:4257",
"http://103.167.170.196:8080",
"http://45.169.98.122:8080",
"http://213.202.216.34:8080",
"http://186.117.148.73:8080",
"http://67.43.236.22:22475",
"http://139.28.232.216:8899",
"http://72.10.164.178:6931",
"http://157.254.53.50:80",
"http://187.94.248.32:15552",
"http://37.187.78.126:3128",
"http://8.218.117.116:1057",
"http://201.47.88.19:3128",
"http://181.209.82.92:1994",
"http://23.82.137.158:80",
"http://38.156.75.60:8080",
"http://217.112.80.252:80",
"http://186.148.195.70:999",
"http://47.247.218.29:3129",
"http://103.187.100.198:3129",
"http://182.253.21.26:46977",
"http://218.1.142.190:57114",
"http://67.43.227.226:27049",
"http://45.70.236.194:999",
"http://142.93.202.130:3128",
"http://43.229.225.210:3128",
"http://47.243.114.192:8180",
"http://182.253.66.122:8080",
"http://178.128.113.118:23128",
"http://103.255.241.101:8080",
"http://201.134.169.214:8204",
"http://146.190.114.113:3128",
"http://46.161.195.182:8080",
"http://103.248.196.73:3125",
"http://67.43.227.226:6345",
"http://103.151.31.21:6969",
"http://64.92.82.58:8080",
"http://79.106.33.26:8079",
"http://61.160.212.169:181",
"http://153.101.67.170:9002",
"http://103.86.109.38:80",
"http://43.252.105.141:8080",
"http://38.156.23.15:999",
"http://200.24.152.210:999",
"http://72.10.164.178:22325",
"http://34.175.101.255:80",
"http://114.218.86.209:8089",
"http://103.200.20.56:3128",
"http://38.183.146.125:8090",
"http://143.42.191.48:80",
"http://188.132.222.10:8080",
"http://41.204.53.24:80",
"http://183.215.23.242:9091",
"http://102.130.125.86:80",
"http://72.10.164.178:31971",
"http://189.22.234.43:80",
"http://38.137.203.13:999",
"http://121.101.133.195:8080",
"http://147.135.128.218:80",
"http://209.14.116.75:999",
"http://60.188.102.225:18080",
"http://118.69.123.180:80",
"http://60.242.169.2:80",
"http://170.254.92.199:30420",
"http://5.161.115.29:51111",
"http://119.252.173.26:8080",
"http://74.48.78.52:80",
"http://182.253.38.179:3128",
"http://45.129.2.216:8080",
"http://5.75.142.128:3128",
"http://183.234.215.11:8443",
"http://87.98.148.98:80",
"http://204.199.174.77:999",
"http://43.134.3.54:13128",
"http://145.40.97.148:9443",
"http://45.92.177.60:8080",
"http://49.156.151.246:83",
"http://45.147.202.252:1234",
"http://27.147.129.26:58080",
"http://72.10.164.178:15765",
"http://202.61.204.51:80",
"http://181.209.125.122:999",
"http://116.125.141.115:80",
"http://169.62.227.172:80",
"http://74.255.219.229:3129",
"http://72.10.160.173:5239",
"http://147.45.76.190:8085",
"http://103.82.8.242:8989",
"http://103.165.212.210:8080",
"http://72.10.160.92:2255",
"http://106.42.30.243:82",
"http://125.26.99.41:8080",
"http://143.198.226.25:80",
"http://91.107.173.67:10514",
"http://67.43.236.19:6613",
"http://8.209.255.13:3128",
"http://154.85.59.71:8443",
"http://189.22.234.36:80",
"http://158.255.77.166:80",
"http://61.14.233.227:1003",
"http://45.224.149.179:999",
"http://41.59.90.174:80",
"http://103.159.194.121:8080",
"http://103.178.13.69:3030",
"http://41.90.123.3:8080",
"http://72.10.160.90:32325",
"http://51.91.109.83:80",
"http://113.108.13.120:4433",
"http://171.251.169.5:4001",
"http://173.249.60.246:14344",
"http://159.203.61.169:8080",
"http://203.128.80.178:8099",
"http://65.108.207.6:80",
"http://108.170.12.10:80",
"http://79.96.240.109:59245",
"http://97.74.87.226:80",
"http://103.162.63.163:8080",
"http://23.82.137.159:80",
"http://45.8.21.156:80",
"http://198.145.118.84:8080",
"http://77.83.246.25:80",
"http://103.154.90.190:8080",
"http://171.228.181.120:26639",
"http://212.224.107.121:80",
"http://185.44.65.171:9595",
"http://45.182.176.38:9947",
"http://188.166.56.246:80",
"http://91.107.196.188:3128",
"http://72.10.160.170:22863",
"http://41.204.53.30:80",
"http://72.10.160.172:1593",
"http://67.43.227.228:17591",
"http://181.78.27.38:999",
"http://80.249.112.164:80",
"http://138.197.102.119:80",
"http://103.157.58.186:8080",
"http://145.40.97.148:10002",
"http://41.60.239.93:8080",
"http://31.220.78.244:80",
"http://110.5.110.158:8080",
"http://67.43.227.228:1167",
"http://65.109.71.86:7777",
"http://149.202.91.219:80",
"http://67.43.228.252:14745",
"http://139.162.78.109:8080",
"http://135.181.154.225:80",
"http://23.95.164.200:80",
"http://113.108.13.120:8083",
"http://45.5.117.168:999",
"http://35.209.198.222:80",
"http://114.218.86.76:8089",
"http://103.180.122.71:8080",
"http://67.43.227.228:3559",
"http://72.10.164.178:28867",
"http://64.38.135.183:80",
"http://43.134.121.40:3128",
"http://93.190.14.205:8080",
"http://31.47.58.37:80",
"http://43.251.84.242:8080",
"http://103.85.114.249:8080",
"http://47.56.110.204:8989",
"http://103.112.66.4:80",
"http://36.64.162.194:8080",
"http://72.10.164.178:14721",
"http://103.234.35.142:8090",
"http://103.231.177.120:5020",
"http://122.54.193.69:8082",
"http://193.150.92.147:80",
"http://165.22.77.86:8080",
"http://145.40.97.148:443",
"http://60.217.33.47:9999",
"http://116.169.61.74:8880",
"http://45.153.165.68:999",
"http://180.190.186.166:8085",
"http://63.143.57.119:80",
"http://223.205.103.1:8080",
"http://62.72.29.174:80",
"http://103.163.227.213:3125",
"http://154.220.66.2:40000",
"http://196.223.129.21:80",
"http://190.14.251.108:999",
"http://67.43.227.227:18573",
"http://47.89.184.18:3128",
"http://67.43.227.227:7813",
"http://67.43.227.226:9797",
"http://36.91.15.241:8080",
"http://177.39.139.14:9999",
"http://203.189.96.232:80",
"http://103.78.158.6:3125",
"http://67.43.236.19:22475",
"http://67.43.227.228:24803",
"http://37.143.129.242:47313",
"http://103.25.79.178:4700",
"http://103.190.179.27:80",
"http://89.221.215.128:80",
"http://159.223.74.131:4750",
"http://72.10.160.172:5903",
"http://182.253.177.198:3128",
"http://72.10.160.91:27473",
"http://202.47.188.193:7777",
"http://129.10.76.179:80",
"http://165.232.129.150:80",
"http://103.82.233.2:53281",
"http://133.18.234.13:80",
"http://72.10.160.90:23463",
"http://189.22.234.39:80",
"http://72.10.160.171:18889",
"http://139.180.164.103:80",
"http://67.43.227.230:10175",
"http://172.191.74.198:8080",
"http://45.77.48.243:80",
"http://181.236.247.8:8815",
"http://78.46.210.178:3128",
"http://91.228.186.200:80",
"http://187.251.224.25:8081",
"http://102.214.164.243:1981",
"http://23.82.137.156:80",
"http://64.147.212.78:8080",
"http://67.43.228.252:3855",
"http://114.130.153.58:58080",
"http://115.74.3.152:10005",
"http://45.133.107.10:81",
"http://148.72.169.225:30119",
"http://178.128.200.87:80",
"http://101.255.119.68:1080",
"http://114.218.86.110:8089",
"http://154.12.242.82:8889",
"http://38.188.127.123:8080",
"http://46.47.197.210:3128",
"http://5.189.158.162:3128",
"http://80.241.251.54:8080",
"http://192.73.244.36:80",
"http://118.179.151.57:5020",
"http://64.227.46.7:8080",
"http://181.236.247.8:8899",
"http://72.10.164.178:18339",
"http://103.189.96.38:8080",
"http://177.73.136.29:8080",
"http://103.147.33.67:7890",
"http://74.208.75.233:3128",
"http://115.72.36.85:10003",
"http://114.218.165.36:8089",
"http://103.254.85.66:8080",
"http://122.10.225.55:8000",
"http://212.110.188.222:34411",
"http://188.125.169.100:8080",
"http://167.99.236.14:80",
"http://87.248.129.26:80",
"http://67.43.228.252:19863",
"http://188.253.112.218:80",
"http://196.204.83.227:1976",
"http://103.139.25.121:8080",
"http://103.78.113.16:8080",
"http://121.227.183.188:8089",
"http://145.40.97.148:10000",
"http://72.10.160.174:8869",
"http://67.43.228.253:5503",
"http://198.74.51.79:8888",
"http://177.70.72.103:20183",
"http://185.118.153.110:8080",
"http://72.10.160.90:15667",
"http://58.136.171.186:8080",
"http://152.230.215.123:80",
"http://31.197.137.186:3128",
"http://189.22.234.32:80",
"http://119.188.93.251:81",
"http://41.254.56.6:1976",
"http://72.10.160.94:1751",
"http://34.95.248.204:3128",
"http://213.233.177.134:80",
"http://58.209.139.136:8089",
"http://203.89.8.107:80",
"http://170.239.205.74:8080",
"http://61.79.169.10:3128",
"http://158.255.77.168:80",
"http://218.252.242.29:80",
"http://67.43.228.250:7567",
"http://157.100.52.178:999",
"http://41.74.91.244:80",
"http://121.227.44.182:8089",
"http://114.130.154.118:58080",
"http://67.43.227.228:15061",
"http://138.118.200.49:999",
"http://103.78.36.116:8080",
"http://190.103.177.131:80",
"http://43.246.200.142:9080",
"http://165.22.77.86:80",
"http://41.203.83.242:8080",
"http://103.148.44.197:8080",
"http://125.212.200.103:9000",
"http://24.152.49.189:999",
"http://72.10.160.171:30393",
"http://37.204.209.191:8080",
"http://201.91.82.155:3128",
"http://67.43.227.228:13085",
"http://37.27.6.46:80",
"http://113.160.132.195:8080",
"http://94.130.177.190:9999",
"http://72.10.164.178:2579",
"http://159.65.221.25:80",
"http://67.43.227.228:26965",
"http://103.170.227.11:82",
"http://145.40.97.148:9401",
"http://47.90.205.231:33333",
"http://67.43.236.20:18301",
"http://43.134.32.184:3128",
"http://200.59.10.50:999",
"http://72.10.164.178:9073",
"http://190.52.104.197:999",
"http://103.199.139.14:83",
"http://43.134.111.70:13128",
"http://106.110.87.245:8081",
"http://103.70.93.77:8080",
"http://38.45.45.60:999",
"http://181.212.45.226:8080",
"http://121.227.179.135:8089",
"http://111.125.242.34:1111",
"http://121.227.179.176:8089",
"http://103.54.80.194:1111",
"http://67.43.236.20:32187",
"http://72.10.164.178:18177",
"http://62.60.236.212:3128",
"http://139.99.239.132:49783",
"http://119.47.90.43:8080",
"http://206.201.3.83:8095",
"http://72.10.160.90:32821",
"http://45.70.201.186:999",
"http://45.117.30.113:58080",
"http://4.157.219.21:80",
"http://103.160.207.49:32650",
"http://84.247.152.41:8080",
"http://58.209.71.87:8089",
"http://177.130.104.106:33333",
"http://50.62.183.223:80",
"http://67.43.228.254:25543",
"http://72.10.160.93:27473",
"http://200.41.170.212:11201",
"http://196.251.195.254:8083",
"http://67.43.227.230:21095",
"http://38.183.146.21:8085",
"http://103.164.117.186:3128",
"http://51.222.102.172:80",
"http://116.98.60.64:10057",
"http://78.28.152.113:80",
"http://121.227.179.187:8089",
"http://194.124.36.106:8080",
"http://60.53.133.218:8080",
"http://62.92.103.241:8080",
"http://202.146.228.251:8088",
"http://128.92.239.190:8080",
"http://103.237.144.232:1311",
"http://183.88.212.184:8080",
"http://103.139.98.175:8080",
"http://43.134.74.143:3128",
"http://200.174.198.86:8888",
"http://180.191.40.160:8082",
"http://177.234.194.226:999",
"http://62.33.207.202:3128",
"http://23.147.112.74:8080",
"http://186.125.218.148:999",
"http://67.43.228.250:16159",
"http://47.243.92.199:3128",
"http://67.43.236.20:30413",
"http://72.10.164.178:31591",
"http://152.228.154.20:80",
"http://45.117.30.129:58080",
"http://51.159.159.73:80",
"http://41.204.53.27:80",
"http://177.19.167.242:80",
"http://113.161.187.190:8080",
"http://198.199.86.11:3128",
"http://45.67.85.64:80",
"http://200.24.130.46:6735",
"http://72.10.160.90:26841",
"http://45.117.29.17:58080",
"http://103.155.197.116:8080",
"http://71.42.125.219:4444",
"http://67.43.236.20:10927",
"http://72.10.160.94:26069",
"http://181.198.64.115:999",
"http://213.169.33.7:4002",
"http://45.189.151.27:999",
"http://42.119.164.44:9400",
"http://61.160.212.99:181",
"http://67.43.227.227:13767",
"http://43.153.207.93:3128",
"http://102.0.5.152:8080",
"http://190.122.88.121:8080",
"http://221.231.13.198:1080",
"http://1.173.149.145:8080",
"http://72.10.160.170:7791",
"http://117.54.114.100:80",
"http://103.169.132.14:8080",
"http://103.123.168.202:3932",
"http://77.119.237.96:8080",
"http://102.68.131.31:8080",
"http://176.236.151.195:50800",
"http://103.133.26.106:8080",
"http://204.157.251.169:999",
"http://136.226.233.109:10742",
"http://51.81.184.60:3128",
"http://41.242.44.9:8084",
"http://194.87.54.124:8082",
"http://5.228.95.209:8081",
"http://185.189.199.75:23500",
"http://154.73.28.79:8080",
"http://103.167.170.62:1111",
"http://51.79.135.181:8080",
"http://222.252.194.29:8080",
"http://199.188.93.55:8000",
"http://103.101.193.78:1111",
"http://103.26.110.37:82",
"http://158.51.210.75:8888",
"http://61.19.109.236:8080",
"http://103.156.141.152:3127",
"http://45.143.108.114:8080",
"http://136.226.255.23:10110",
"http://103.160.184.204:8080",
"http://103.176.96.87:8085",
"http://101.255.209.242:8080",
"http://103.124.173.209:82",
"http://191.102.254.11:8085",
"http://191.97.68.42:8080",
"http://38.156.73.61:8080",
"http://102.214.104.56:8080",
"http://103.48.68.124:83",
"http://188.240.120.15:8080",
"http://27.254.46.194:80",
"http://200.94.102.12:999",
"http://222.252.194.204:8080",
"http://148.72.168.81:30118",
"http://191.102.248.9:8085",
"http://94.102.234.186:32650",
"http://103.118.175.154:8080",
"http://103.29.185.59:3125",
"http://103.189.197.113:8181",
"http://203.190.44.251:8080",
"http://103.80.230.26:8090",
"http://167.103.3.22:11011",
"http://190.95.202.210:999"
]

import time
import random
from selenium import webdriver
from selenium.webdriver.common.by import By
from tqdm import tqdm
import pandas as pd
from fake_useragent import UserAgent
from bs4 import BeautifulSoup
ua = UserAgent()

df = pd.DataFrame(columns=["Title", "Description", "URL"])
options = webdriver.ChromeOptions()
options.add_argument("--disable-gpu")
options.add_argument(f"user-agent={ua.random}")
driver = webdriver.Chrome(options=options)
driver.implicitly_wait(10)

def extract_content_from_page(driver):
    html_content = driver.page_source
    print(html_content)
    soup = BeautifulSoup(html_content, 'html.parser')
    
    title = soup.find('h1')
    description = soup.find('meta', {'name': 'description'})
    url = driver.current_url
    
    title_text = title.get_text(strip=True) if title else "No Title Found"
    description_text = description['content'] if description else "No Description Found"
    
    return title_text, description_text, url

for index, link in enumerate(tqdm(generated_links, desc="Processing Links", unit="link")):
    if index > 0 and index % 5 == 0:
        driver.quit()
        proxy = random.choice(proxy_list)
        options = webdriver.ChromeOptions()
        options.add_argument(f"--proxy-server={proxy}")
        options.add_argument("--disable-gpu")
        options.add_argument(f"user-agent={ua.random}")
        driver = webdriver.Chrome(options=options)
        driver.implicitly_wait(10)
    
    try:

        driver.get(link)

        time.sleep(5)
        
        title, description, url = extract_content_from_page(driver)
        
        new_data = pd.DataFrame([{"Title": title, "Description": description, "URL": url}])
        
        df = pd.concat([df, new_data], ignore_index=True)
        
    except Exception as e:
        print(f"Error processing {link} with proxy {proxy}")
        continue
driver.quit()
print(df)
