==> 2 Preprocessing.ipynb <==
### Imports

import pandas as pd
import nltk
from nltk.corpus import stopwords
from nltk.stem import WordNetLemmatizer
from tqdm import tqdm

tqdm.pandas()
nltk.download('stopwords')
nltk.download('wordnet')

### Load articles

article_df = pd.read_csv("bbc_full.csv").astype(str)
display(article_df.head())

import numpy as np 
np.mean(list(map(lambda x: len(x), article_df["content"])))

### Preprocess

stop_words = set(stopwords.words("english"))
lemmatizer = WordNetLemmatizer()

def preprocess_text(text):
    # # 1. Lowercase the text
    # text = text.lower()
    
    # # 2. Remove stopwords
    # text = " ".join(word for word in text.split() if word not in stop_words)
    
    # # 3. Remove non-alphanumeric characters
    # text = " ".join(word for word in text.split() if word.isalpha())
    
    # # 4. Lemmatization
    # text = " ".join(lemmatizer.lemmatize(word) for word in text.split())

    text = " ".join(lemmatizer.lemmatize(word) for word in text.lower().split() if word.isalpha() and word not in stop_words)
    
    return text

# Apply preprocessing to the specified columns
article_df['title_cleaned'] = article_df['title'].progress_apply(preprocess_text)
article_df['description_cleaned'] = article_df['description'].progress_apply(preprocess_text)
article_df['content_cleaned'] = article_df['content'].progress_apply(preprocess_text)

display(article_df.head())

print(f"{len(article_df['section'].unique())} categories found")
display(sorted(article_df["section"].unique()))

### Articles filtering

tech_categories = [
    'BBC Trending',
    'BBC Verify',
    'Business',
    'Companies',
    'Climate',
    'Education & Family',
    'Features',
    'Health',
    'Inside Europe Blog',
    'Long Reads',
    'Reality Check',
    'Science & Environment',
    'Sustainability',
    'Technology',
    'UK Politics',
    'US & Canada',
    'World',
    'Your Money'
]

article_df["section"] = article_df["section"].progress_apply(lambda x: x if x in tech_categories else "nan")
article_df = article_df[article_df["section"] != "nan"]
display(article_df)

### Save preprocessed data

columns = ["published_date","section",'title_cleaned',"description_cleaned","content_cleaned"]
processed_data = article_df[columns]
processed_data.columns = ["date","section","title","description","content"]
processed_data["date"] = pd.to_datetime(processed_data["date"])

display(processed_data.head())

processed_data.to_csv("bbc_preprocessed.csv", index=False)
